Avoid linking AO aux tables to extension (#930)
Unlike TOAST tables, AO aux tables are linked to extension during its creation.
When table is dropped from extension using ALTER, aux tables are left as
extension dependencies. It means we can't DROP a table after that, because AO
aux tables are still used by extension. In other words, this behavior makes
impossible to DROP AO table that was created by extension without extension
DROP.

The goal of the patch is to change this behavior, allowing to DROP a table.
Unnecessary dependencies are now not created in heap_create_with_catalog(). The
behavior is the same as for TOAST tables, which are aux tables too. The
dependency created in CreateAOAuxiliaryTable() is enough.

The patch contains a test that describes target behavior.

This is a backport of dfe49e3. The test was
changed as 6x has no test_ext* extensions. To achieve the same result, new test
extension was made.

(cherry picked from commit 10ec9c7ae00475f0db823d53a038b3768b679d63)