Throw error if one has a few refs to modifying CTE
Multiple references to CTEs which contain modifying DML operations can lead
to duplicated data modifications. This commit prohibits such cases and cuts
the query planning with the informative error. The if-clause, which checks
whether current cte is referenced more than once and contains any subquery
with any non-SELECT operation, is proposed. This solution required introduction
of a special walker that covers various cases of subqueries.